remove build tooling, remove bluebird, adjust for node conventions, not es6 modules (for now)
