+++ lib/api/getSize.js
@@ -1,2 +1,18 @@
-'use strict';Object.defineProperty(exports,'__esModule',{value:true});var _toolsSmb2Forge=require('../tools/smb2-forge');exports['default'] = function(path,cb){(0,_toolsSmb2Forge.request)('open',{path:path},this,function(err,file){if(err){if(err.code === 'STATUS_OBJECT_NAME_NOT_FOUND'){err.code = 'ENOENT';}cb(err);}else {var fileLength=0;for(var i=0;i < file.EndofFile.length;i++) {fileLength += file.EndofFile[i] * Math.pow(2,i * 8);}cb(null,fileLength);}});};module.exports = exports['default'];
-//# sourceMappingURL=getSize.js.map
+const {request} = require('../tools/smb2-forge');
+
+module.exports = (path, cb) => {
+  request('open', {path}, this, (err, file) => {
+    if (err) {
+      if (err.code === 'STATUS_OBJECT_NAME_NOT_FOUND') {
+        err.code = 'ENOENT'
+      }
+      cb(err)
+    } else {
+      let fileLength = 0
+      for (let i = 0; i < file.EndofFile.length; i++) {
+        fileLength += file.EndofFile[i] * Math.pow(2, i * 8)
+      }
+      cb(null, fileLength)
+    }
+  })
+}
