+++ lib/structures/constants.js
@@ -8,5 +8,9 @@ module.exports = {
   FILE_CREATE: 0x00000002,
   FILE_OPEN_IF: 0x00000003,
   FILE_OVERWRITE: 0x00000004,
-  FILE_OVERWRITE_IF: 0x00000005
+  FILE_OVERWRITE_IF: 0x00000005,
+  FILE_SHARE_NONE: 0x00000000,
+  FILE_SHARE_READ: 0x00000001,
+  FILE_SHARE_WRITE: 0x00000002,
+  FILE_SHARE_DELETE: 0x00000004
 }
